Factored common http code into base class
