+++ lib/smess.rb
@@ -11,6 +11,7 @@ require 'smess/logging'
 require 'smess/country_code_registry'
 require 'smess/utils'
 require 'smess/sms'
+require 'smess/outputs/http_base'
 require 'smess/outputs/auto'
 require 'smess/outputs/ipx'
 require 'smess/outputs/ipxus'
